requirements.txt
# FastAPI framework
fastapi==0.110.0

# ASGI server for running FastAPI
uvicorn[standard]==0.27.1

# MongoDB async driver
motor==3.7.0

# Redis async driver
redis==5.0.1

# Pydantic for data validation
pydantic==2.5.3

# Password hashing & authentication
passlib[bcrypt]==1.7.4

# JWT Authentication
pyjwt==2.8.0

# Load environment variables from .env
python-dotenv==1.0.1

# Logging
loguru==0.7.2
